Running dengue abc using sql; new seropres metric
Everything seems to be working, although there is likely clean up to do.

We're now storing the seeds using for the simulator's RNG, so that results
are actually repeatable.

We've added another metric, the fraction of 8-14 year olds we expect to be
seropositive in the Merida population.  This creates a bit of ugliness in the
simulation interface, since that's not a generic thing to report.

+++ Parameters.h
@@ -176,7 +176,7 @@ public:
     void generateAnnualSerotypes();
     bool simulateAnnualSerotypes;
 
-    int randomseed;
+    unsigned long int randomseed;
     int nRunLength;
     double betaPM;                                          // scales person-to-mosquito transmission
     double betaMP;                                          // scales mosquito-to-person transmission (includes bite rate)
